refactor(models): leagues.dart에서 league.dart로 파일명 수정
